refactor : 상품 목록 API 커스텀 훅 생성
