+++ src/components/search/SearchProducts.tsx
@@ -1,27 +1,15 @@
 'use client';
 
-import { useProductQueryStore } from '@/stores/productQuery';
 import ProductList from '../products/ProductList';
 import SortOptions from '../subHeader/SortOptions';
 import NullProductList from './NullProductList';
-import { keepPreviousData, useInfiniteQuery } from '@tanstack/react-query';
-import { getProductsFetcher } from '@/fetchers/product';
 import { Product } from '@/models/product';
 import Skeleton from '../products/Skeleton';
+import useProducts from '@/hooks/product';
 
 export default function SearchProducts() {
-  const productQuery = useProductQueryStore.use.productQuery();
   const { data, isLoading, fetchNextPage, hasNextPage, isFetching } =
-    useInfiniteQuery({
-      queryKey: ['products', productQuery],
-      queryFn: ({ pageParam }) =>
-        getProductsFetcher({ cursor: pageParam, query: productQuery }),
-      initialPageParam: null,
-      getNextPageParam: (lastPage) => lastPage.cursor,
-      select: (data) => data.pages,
-      placeholderData: keepPreviousData,
-      refetchOnWindowFocus: false,
-    });
+    useProducts();
   const products: Product[] = data?.map((page) => page.results).flat() || [];
 
   if (isLoading) return <Skeleton />;
