+++ src/hooks/product.tsx
@@ -1,8 +1,33 @@
-import { addFavoriteProductFetcher } from '@/fetchers/product';
+import {
+  addFavoriteProductFetcher,
+  getProductsFetcher,
+} from '@/fetchers/product';
 import { useProductQueryStore } from '@/stores/productQuery';
 import { useToastStore } from '@/stores/toast';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import {
+  keepPreviousData,
+  useInfiniteQuery,
+  useMutation,
+  useQueryClient,
+} from '@tanstack/react-query';
 import axios from 'axios';
+
+export default function useProducts() {
+  const productQuery = useProductQueryStore.use.productQuery();
+  const { data, isLoading, fetchNextPage, hasNextPage, isFetching } =
+    useInfiniteQuery({
+      queryKey: ['products', productQuery],
+      queryFn: ({ pageParam }) =>
+        getProductsFetcher({ cursor: pageParam, query: productQuery }),
+      initialPageParam: null,
+      getNextPageParam: (lastPage) => lastPage.cursor,
+      select: (data) => data.pages,
+      placeholderData: keepPreviousData,
+      refetchOnWindowFocus: false,
+    });
+
+  return { data, isLoading, fetchNextPage, hasNextPage, isFetching };
+}
 
 export function useAddFavorite() {
   const queryClient = useQueryClient();
